Create new file structure json files should be saved in json folder. schema sql files should be saved inside schemafolder unit tests inside spec folder. moved our classes to the classes folder and module files inside the modules folder.
