+++ classes/author.rb
@@ -1,4 +1,4 @@
-require_relative './item'
+require_relative 'item'
 
 class Author < Item
   attr_accessor :first_name, :last_name
