+++ classes/game.rb
@@ -1,4 +1,4 @@
-require_relative './item'
+require_relative 'item'
 
 class Game < Item
   attr_accessor :multiplayer, :cover_state
