+++ classes/label.rb
@@ -1,4 +1,4 @@
-require './item'
+require_relative 'item'
 
 # Create label class
 class Label
